reorganized CampaignCreator folder. Files renamed and old triggeredeventsSettings file broken up.
