+++ client/src/components/CampaignCreator/MonstersAndModules/RadioButtonListCampaignCreator.tsx
@@ -1,7 +1,7 @@
 import { Flex } from '@radix-ui/themes';
-import { TwilightNodeHeader, TwilightRadioGroupItem } from '../primitiveComponents/Primitives';
-import { campaignCreatorData } from '../static_data_file_configs/CampaignCreatorConfig';
-import { NodeCoreFinaleLists, TypeCampaignData } from '../static_data_file_configs/PresetCampaignConfig';
+import { TwilightNodeHeader, TwilightRadioGroupItem } from '../../primitiveComponents/Primitives';
+import { campaignCreatorData } from '../../static_data_file_configs/CampaignCreatorConfig';
+import { NodeCoreFinaleLists, TypeCampaignData } from '../../static_data_file_configs/PresetCampaignConfig';
 
 interface RadioButtonListCampaignCreatorProps {
     header: string;
