+++ client/src/components/CampaignCreator/TimelineCreator/index.tsx
@@ -1,8 +1,8 @@
 import { Button, Flex, Heading, Separator } from '@radix-ui/themes';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import TimelineTable from '../primitiveComponents/TimelineTable';
-import { TypeCampaignData } from '../static_data_file_configs/PresetCampaignConfig';
+import TimelineTable from '../../primitiveComponents/TimelineTable';
+import { TypeCampaignData } from '../../static_data_file_configs/PresetCampaignConfig';
 
 interface TimelineCreatorProps {
     campaignSettings: TypeCampaignData;
